pyproject.toml
[project]
name = "v-raptor"
version = "0.1.0"
dependencies = [
    "alembic==1.17.1",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "cachetools==6.2.1",
    "certifi==2025.10.5",
    "charset-normalizer==3.4.4",
    "diskcache==5.6.3",
    "docker==7.1.0",
    "google-api-core==2.28.1",
    "google-api-python-client==2.186.0",
    "google-auth==2.42.1",
    "google-auth-httplib2==0.2.1",
    "google-genai==0.3.0",
    "googleapis-common-protos==1.71.0",
    "grpcio==1.76.0",
    "grpcio-status==1.71.2",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.31.0",
    "httpx",
    "idna==3.11",
    "Jinja2==3.1.6",
    "llama-cpp-python==0.3.16",
    "MarkupSafe==3.0.3",
    "numpy>2.2",
    "proto-plus==1.26.1",
    "protobuf==5.29.5",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.12.3",
    "pydantic_core==2.41.4",
    "PyJWT==2.8.0",
    "pyparsing==3.2.5",
    "requests==2.32.5",
    "rsa==4.9.1",
    "sniffio==1.3.1",
    "SQLAlchemy==2.0.31",
    "tenacity==9.1.2",
    "tqdm==4.67.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "websockets==14.2",
    "GitPython==3.1.43",
    "Flask==3.0.3",
    "redis==5.0.7",
    "rq==1.16.2",
    "PyGithub==2.5.0",
    "ollama",
    "radon",
    "googlesearch-python",
]
